+++ CHANGELOG.md
@@ -76,6 +76,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/cvat-ai/datumaro/pull/9>, <https://github.com/cvat-ai/datumaro/pull/16>)
 - Missing comparison of the base class attributes in the `Mask` class
   (<https://github.com/cvat-ai/datumaro/pull/28>)
+- Image stats when no image info available for some images in the dataset
+  (<https://github.com/cvat-ai/datumaro/pull/29>)
 
 ### Security
 - TBD